backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-text-splitters==0.3.4
langchain-qdrant==0.2.0
qdrant-client==1.12.1
fastembed==0.4.2
pymupdf==1.25.3
python-docx==1.1.2
sentence-transformers==3.3.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
tiktoken==0.8.0
aiofiles==24.1.0
